+++ apps/modernjs-ssr/dynamic-nested-remote/@mf-types/dynamic_remote/compiled-types/Image.d.ts
@@ -1,0 +1,4 @@
+declare const _default: ({ text }: {
+    text: string;
+}) => JSX.Element;
+export default _default;