+++ apps/router-demo/router-remote1-2001/src/export-App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import App from './App';
 import { createBridgeComponent } from '@module-federation/bridge-react';
 
-export const provider = createBridgeComponent({
+const provider = createBridgeComponent({
   rootComponent: App,
 });