+++ apps/router-demo/router-remote4-2004/src/export-App.tsx
@@ -1,0 +1,8 @@
+import App from './App';
+import { createBridgeComponent } from '@module-federation/bridge-react';
+
+const provider = createBridgeComponent({
+  rootComponent: App,
+});
+
+export default provider;