+++ packages/enhanced/src/index.ts
@@ -1,3 +1,4 @@
+import type { moduleFederationPlugin } from '@module-federation/sdk';
 export { default as ModuleFederationPlugin } from './wrapper/ModuleFederationPlugin';
 export { default as ContainerReferencePlugin } from './wrapper/ContainerReferencePlugin';
 export { default as SharePlugin } from './wrapper/SharePlugin';
@@ -9,16 +10,16 @@ export { default as FederationRuntimePlugin } from './wrapper/FederationRuntimeP
 export { default as AsyncBoundaryPlugin } from './wrapper/AsyncBoundaryPlugin';
 export { default as HoistContainerReferencesPlugin } from './wrapper/HoistContainerReferencesPlugin';
 
-export {
-  isRequiredVersion,
-  normalizeVersion,
-  getDescriptionFile,
-  getRequiredVersionFromDescriptionFile,
-} from './lib/sharing/utils';
 export { parseOptions } from './lib/container/options';
 
 export const container = {
   get ContainerEntryModule() {
     return require('./lib/container/ContainerEntryModule').default;
   },
+};
+
+export const createModuleFederationConfig = (
+  options: moduleFederationPlugin.ModuleFederationPluginOptions,
+): moduleFederationPlugin.ModuleFederationPluginOptions => {
+  return options;
 };