+++ packages/runtime-tools/src/runtime.ts
@@ -1,1 +1,2 @@
 export * from '@module-federation/runtime';
+export { default as RetryPlugin } from '@module-federation/runtime/retry-plugin';