+++ packages/runtime/src/index.ts
@@ -91,5 +91,9 @@ export function registerPlugins(
   return FederationInstance.registerPlugins.apply(FederationInstance, args);
 }
 
+export function getInstance() {
+  return FederationInstance;
+}
+
 // Inject for debug
 setGlobalFederationConstructor(FederationHost);