+++ packages/runtime/src/plugins/generate-preload-assets.ts
@@ -22,6 +22,7 @@ import { getRegisteredShare } from '../utils/share';
 import {
   arrayOptions,
   getFMId,
+  getRemoteEntryInfoFromSnapshot,
   isPureRemoteEntry,
   isRemoteInfoWithEntry,
 } from '../utils';
@@ -149,9 +150,7 @@ export function generatePreloadAssets(
 
       const remoteEntryUrl = getResourceUrl(
         moduleInfoSnapshot,
-        'remoteEntry' in moduleInfoSnapshot
-          ? moduleInfoSnapshot.remoteEntry
-          : '',
+        getRemoteEntryInfoFromSnapshot(moduleInfoSnapshot).url,
       );
 
       if (remoteEntryUrl) {