+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @module-federation/bridge-react
 
+## 0.5.1
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- 49d6135: feat(@module-federation/bridge): enhance Bridge capabilities and fix some issues
+- Updated dependencies [49d6135]
+  - @module-federation/bridge-shared@0.5.0
+
+## 0.4.0
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.4.0
+
+## 0.3.5
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.5
+
+## 0.3.4
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.4
+
+## 0.3.3
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.3
+
+## 0.3.2
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.2
+
+## 0.3.1
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.1
+
+## 0.3.0
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.3.0
+
+## 0.2.8
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.2.8
+
 ## 0.2.7
 
 ### Patch Changes