+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @module-federation/bridge-shared
 
+## 0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- 49d6135: feat(@module-federation/bridge): enhance Bridge capabilities and fix some issues
+
+## 0.4.0
+
+## 0.3.5
+
+## 0.3.4
+
+## 0.3.3
+
+## 0.3.2
+
+## 0.3.1
+
+## 0.3.0
+
+## 0.2.8
+
 ## 0.2.7
 
 ## 0.2.6