+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @module-federation/esbuild
 
+## 0.0.16
+
+### Patch Changes
+
+- @module-federation/sdk@0.5.1
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [8378a77]
+  - @module-federation/sdk@0.5.0
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [a6e2bed]
+- Updated dependencies [a6e2bed]
+  - @module-federation/sdk@0.4.0
+
+## 0.0.13
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.5
+
+## 0.0.12
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.4
+
+## 0.0.11
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.3
+
+## 0.0.10
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.2
+
+## 0.0.9
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.1
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [fa37cc4]
+  - @module-federation/sdk@0.3.0
+
+## 0.0.7
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.8
+
 ## 0.0.6
 
 ### Patch Changes