+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @module-federation/runtime
 
+## 0.5.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.5.1
+
+## 0.5.0
+
+### Minor Changes
+
+- 8378a77: feat(runtime): add loadEntry hook
+
+### Patch Changes
+
+- 5c7ac8a: provide error message when remote entry init fails
+- Updated dependencies [8378a77]
+  - @module-federation/sdk@0.5.0
+
+## 0.4.0
+
+### Minor Changes
+
+- f2f02c9: provide users to mark shared module as loaded
+
+### Patch Changes
+
+- a6e2bed: fix: add attrs option to createLink hook
+- Updated dependencies [a6e2bed]
+- Updated dependencies [a6e2bed]
+  - @module-federation/sdk@0.4.0
+
+## 0.3.5
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.5
+
+## 0.3.4
+
+### Patch Changes
+
+- 67ea678: checking for remote entry type when loading assets
+- d26d7e6: Ensured createScript runtime hook always receives `attrs`
+  - @module-federation/sdk@0.3.4
+
+## 0.3.3
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.3
+
+## 0.3.2
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.2
+
+## 0.3.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.1
+
+## 0.3.0
+
+### Minor Changes
+
+- fa37cc4: feat: support modern.js ssr [#2348](https://github.com/module-federation/core/issues/2348)
+
+### Patch Changes
+
+- Updated dependencies [fa37cc4]
+  - @module-federation/sdk@0.3.0
+
+## 0.2.8
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.8
+
 ## 0.2.7
 
 ### Patch Changes