+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.5.1
+
+## 0.5.0
+
+## 0.4.0
+
+## 0.3.5
+
+## 0.3.4
+
+## 0.3.3
+
+## 0.3.2
+
+### Patch Changes
+
+- 0de1c83: fix: do not collect node internal utils
+
+## 0.3.1
+
+## 0.3.0
+
+## 0.2.8
+
 ## 0.2.7
 
 ### Patch Changes